Move sourcemap tests to `cli-integration`
